added async lock
